requirements.txt
streamlit
langchain
langchain-core
langgraph
langchain-groq
pydantic<3,>=2
pandas
plotly
python-dotenv
httpx
alpha_vantage
fastapi
uvicorn[standard]
tavily
langchain_ollama
langchain-google-genai